chore: allow for longer announcement and maintenance messages
Take advantage of the new `max_length` param to allow for longer messages. Not useful for purely text but helpful if they want to set HTML code.

e.g. some sellers have wanted to add links in the announcement text banner which is nearly impossible with markup and still <= 100 chars.